:truck: move `find-tutorial` to `src/`
