+++ api/src/devcomp/domain/usecases/find-tutorials.js
@@ -1,8 +1,9 @@
 import _ from 'lodash';
 
-import { Scorecard } from '../../../src/evaluation/domain/models/Scorecard.js';
-import { UserNotAuthorizedToAccessEntityError } from '../../../src/shared/domain/errors.js';
-import { KnowledgeElement } from '../../../src/shared/domain/models/KnowledgeElement.js';
+// TODO Bounded context violation
+import { Scorecard } from '../../../evaluation/domain/models/Scorecard.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../shared/domain/errors.js';
+import { KnowledgeElement } from '../../../shared/domain/models/KnowledgeElement.js';
 
 const findTutorials = async function ({
   authenticatedUserId,
@@ -13,8 +14,10 @@ const findTutorials = async function ({
   tutorialRepository,
   locale,
 }) {
+  // TODO replace scorecardId by competenceId
   const { userId, competenceId } = Scorecard.parseId(scorecardId);
 
+  // TODO refactor to extract this guard to controller
   if (parseInt(authenticatedUserId) !== parseInt(userId)) {
     throw new UserNotAuthorizedToAccessEntityError();
   }
